// interfaces/cooker/cooker.interface.ts
export interface Cooker {
  id: number
  name: string
  specialty: string
  rating: number
  reviews: number
  location: string
  price: string
  tag: string
  from: string
  to: string
  initials: string
  subscribers: number
  meals: number
  plan: string
  planDetail: string
  bio: string
  foodImage: string
}